requirements.txt
# Auto-generated on 2024-12-18 13:51:09
Flask==3.1.0
Flask-Assets==2.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Jinja2==3.1.4
Mako==1.3.7
MarkupSafe==3.0.2
Pygments==2.18.0
SQLAlchemy==2.0.36
WTForms==3.2.1
Werkzeug==3.1.3
alembic==1.14.0
blinker==1.9.0
click==8.1.7
itsdangerous==2.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
pillow==11.0.0
pip==24.3.1
python-dotenv==1.0.1
python-webp==2.0.1
rich==13.9.4
setuptools==75.6.0
shellingham==1.5.4
typer==0.15.1
typing_extensions==4.12.2
webassets==2.0
